extracted WeightInfoExt into separate file
