+++ modules/message-lane/src/lib.rs
@@ -35,6 +35,8 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub use crate::weights_ext::{ensure_weights_are_correct, WeightInfoExt};
+
 use crate::inbound_lane::{InboundLane, InboundLaneStorage};
 use crate::outbound_lane::{OutboundLane, OutboundLaneStorage};
 use crate::weights::WeightInfo;
@@ -59,6 +61,7 @@ use sp_std::{cell::RefCell, marker::PhantomData, prelude::*};
 
 mod inbound_lane;
 mod outbound_lane;
+mod weights_ext;
 
 pub mod instant_payments;
 pub mod weights;
@@ -76,7 +79,7 @@ pub trait Config<I = DefaultInstance>: frame_system::Config {
 	/// They overarching event type.
 	type Event: From<Event<Self, I>> + Into<<Self as frame_system::Config>::Event>;
 	/// Benchmarks results from runtime we're plugged into.
-	type WeightInfo: WeightInfoEx;
+	type WeightInfo: WeightInfoExt;
 	/// Maximal number of messages that may be pruned during maintenance. Maintenance occurs
 	/// whenever new message is sent. The reason is that if you want to use lane, you should
 	/// be ready to pay for its maintenance.
@@ -133,66 +136,6 @@ pub trait Config<I = DefaultInstance>: frame_system::Config {
 	/// Message dispatch.
 	type MessageDispatch: MessageDispatch<Self::InboundMessageFee, DispatchPayload = Self::InboundPayload>;
 }
-
-/// Extended weight info.
-pub trait WeightInfoEx: WeightInfo {
-	/// Returns weight overhead of message delivery transaction (`receive_messages_proof`).
-	fn receive_messages_proof_overhead() -> Weight {
-		let weight_of_two_messages_and_two_tx_overheads = Self::receive_single_message_proof().saturating_mul(2);
-		let weight_of_two_messages_and_single_tx_overhead = Self::receive_two_messages_proof();
-		weight_of_two_messages_and_two_tx_overheads.saturating_sub(weight_of_two_messages_and_single_tx_overhead)
-	}
-
-	/// Returns weight that needs to be accounted when receiving given number of messages with message
-	/// delivery transaction (`receive_messages_proof`).
-	fn receive_messages_proof_messages_overhead(messages: MessageNonce) -> Weight {
-		let weight_of_two_messages_and_single_tx_overhead = Self::receive_two_messages_proof();
-		let weight_of_single_message_and_single_tx_overhead = Self::receive_single_message_proof();
-		weight_of_two_messages_and_single_tx_overhead
-			.saturating_sub(weight_of_single_message_and_single_tx_overhead)
-			.saturating_mul(messages as Weight)
-	}
-
-	/// Returns weight that needs to be accounted when message delivery transaction (`receive_messages_proof`)
-	/// is carrying outbound lane state proof.
-	fn receive_messages_proof_outbound_lane_state_overhead() -> Weight {
-		let weight_of_single_message_and_lane_state = Self::receive_single_message_proof_with_outbound_lane_state();
-		let weight_of_single_message = Self::receive_single_message_proof();
-		weight_of_single_message_and_lane_state.saturating_sub(weight_of_single_message)
-	}
-
-	/// Returns weight overhead of delivery confirmation transaction (`receive_messages_delivery_proof`).
-	fn receive_messages_delivery_proof_overhead() -> Weight {
-		let weight_of_two_messages_and_two_tx_overheads =
-			Self::receive_delivery_proof_for_single_message().saturating_mul(2);
-		let weight_of_two_messages_and_single_tx_overhead =
-			Self::receive_delivery_proof_for_two_messages_by_single_relayer();
-		weight_of_two_messages_and_two_tx_overheads.saturating_sub(weight_of_two_messages_and_single_tx_overhead)
-	}
-
-	/// Returns weight that needs to be accounted when receiving confirmations for given number of
-	/// messages with delivery confirmation transaction (`receive_messages_delivery_proof`).
-	fn receive_messages_delivery_proof_messages_overhead(messages: MessageNonce) -> Weight {
-		let weight_of_two_messages = Self::receive_delivery_proof_for_two_messages_by_single_relayer();
-		let weight_of_single_message = Self::receive_delivery_proof_for_single_message();
-		weight_of_two_messages
-			.saturating_sub(weight_of_single_message)
-			.saturating_mul(messages as Weight)
-	}
-
-	/// Returns weight that needs to be accounted when receiving confirmations for given number of
-	/// relayers entries with delivery confirmation transaction (`receive_messages_delivery_proof`).
-	fn receive_messages_delivery_proof_relayers_overhead(relayers: MessageNonce) -> Weight {
-		let weight_of_two_messages_by_two_relayers = Self::receive_delivery_proof_for_two_messages_by_two_relayers();
-		let weight_of_two_messages_by_single_relayer =
-			Self::receive_delivery_proof_for_two_messages_by_single_relayer();
-		weight_of_two_messages_by_two_relayers
-			.saturating_sub(weight_of_two_messages_by_single_relayer)
-			.saturating_mul(relayers as Weight)
-	}
-}
-
-impl<T: WeightInfo> WeightInfoEx for T {}
 
 /// Shortcut to messages proof type for Config.
 type MessagesProofOf<T, I> =
@@ -595,19 +538,6 @@ impl<T: Config<I>, I: Instance> Module<T, I> {
 			total_messages: total_unrewarded_messages(&relayers),
 		}
 	}
-}
-
-/// Ensure that weights from `WeightInfo` implementation are looking correct.
-pub fn ensure_weights_are_correct<W: WeightInfoEx>() {
-	assert_ne!(W::send_message_worst_case(), 0);
-
-	assert_ne!(W::receive_messages_proof_overhead(), 0);
-	assert_ne!(W::receive_messages_proof_messages_overhead(1), 0);
-	assert_ne!(W::receive_messages_proof_outbound_lane_state_overhead(), 0);
-
-	assert_ne!(W::receive_messages_delivery_proof_overhead(), 0);
-	assert_ne!(W::receive_messages_delivery_proof_messages_overhead(1), 0);
-	assert_ne!(W::receive_messages_delivery_proof_relayers_overhead(1), 0);
 }
 
 /// Getting storage keys for messages and lanes states. These keys are normally used when building
